+++ public/main.css
@@ -13,7 +13,7 @@
 
 /** Sign in page **/
 
-@import "components/signin/_signin.css";
+@import "components/signin-form/_signin-form.css";
 @import "components/oauth-cta/_oauth-cta.css";
 
 /** MVT: Sign in v2 variant B **/
