+++ public/signin-page.hbs
@@ -13,47 +13,7 @@
 
   {{> components/oauth-cta/_oauth-cta }}
 
-  <form id="signin_form" class="signin-form" action="{{ actions.signIn }}" method="POST">
-    <input type="hidden" name="{{ csrfToken.fieldName }}" value="{{ csrfToken.value }}">
-    <input type="hidden" name="returnUrl" value="{{returnUrl}}">
-    <input type="hidden" name="skipConfirmation" value="{{skipConfirmation}}">
-    <p class="signin-form__prelude">{{signInPageText.divideText}}</p>
-
-    <fieldset class="signin-form__fieldset">
-
-      {{#each errors}}
-          <p id="{{id}}" class="signin-form__error">{{message}}</p>
-      {{/each}}
-
-      <div class="signin-form__control--email">
-        <label class="signin-form__label--email" for="signin_field_email">{{signInPageText.email}}</label>
-        <input class="signin-form__field--email" id="signin_field_email" type="email" name="email" autocomplete="on" autocapitalize="off" autocorrect="off" spellcheck="false" value="{{email}}"/>
-      </div>
-      <div class="signin-form__control--password">
-        <label class="signin-form__label--password" for="signin_field_password">{{signInPageText.password}}</label>
-        <input class="signin-form__field--password" id="signin_field_password" type="password" name="password" />
-      </div>
-      <div class="signin-form__cta--reset"><a href="{{forgotPasswordUrl}}">{{signInPageText.forgottenPassword}}</a></div>
-
-      <div class="signin-form__control--remember-me">
-        <input class="signin-form__checkbox--remember-me" id="signin_field_remember_me" type="checkbox" name="rememberMe" checked="checked" value="true"/>
-        <label class="signin-form__label--remember-me" for="signin_field_remember_me">{{signInPageText.rememberMe}}</label>
-      </div>
-
-      {{#if recaptchaModel}}
-        <div class="signin-form__recaptcha-wrapper">
-          <div class="g-recaptcha signin-form__recaptcha" data-sitekey="{{recaptchaModel.googleSiteKey}}"></div>
-        </div>
-      {{/if}}
-
-      <div class="signin-form__control--submit">
-        <button class="signin-form__submit-button" id="signin_submit" type="submit">{{signInPageText.signIn}}</button>
-      </div>
-
-      <p class="signin-form__cta--signup"> {{signInPageText.noAccount}} <a id="register_cta" href="{{registerUrl}}" > {{signInPageText.signUp}} </a>.</p>
-
-    </fieldset>
-  </form>
+  {{> components/signin-form/_signin-form }}
 
   {{> components/terms/_terms }}
